+++ lib/calendav/clients/events_client.rb
@@ -13,11 +13,28 @@ module Calendav
       end
 
       def create(calendar_url, event_identifier, ics)
-        update merged_url(calendar_url, event_identifier), ics
+        event_url = merged_url(calendar_url, event_identifier)
+        result = endpoint.put(ics, url: event_url, content_type: :ics)
+
+        result.headers["Location"] || event_url
       end
 
-      def delete(event_url)
-        endpoint.delete(url: event_url)
+      def delete(event_url, etag: nil)
+        endpoint.delete(url: event_url, etag: etag)
+      rescue Endpoint::RequestFailedError => error
+        return false if error.message["412"]
+
+        raise
+      end
+
+      def find(event_url)
+        response = endpoint.get(url: event_url)
+
+        Event.new(
+          event_url,
+          calendar_data: response.body.to_s,
+          etag: response.headers["ETag"]
+        )
       end
 
       def list(calendar_url, from: nil, to: nil)
@@ -30,10 +47,16 @@ module Calendav
           .collect { |node| Event.from_xml(calendar_url, node) }
       end
 
-      def update(event_url, ics)
-        result = endpoint.put(ics, url: event_url, content_type: :ics)
+      def update(event_url, ics, etag: nil)
+        endpoint.put(
+          ics, url: event_url, content_type: :ics, etag: etag
+        )
 
-        result.headers["Location"] || event_url
+        true
+      rescue Endpoint::RequestFailedError => error
+        return false if error.message["412"]
+
+        raise
       end
 
       private