+++ lib/calendav/event.rb
@@ -1,32 +1,23 @@
 # frozen_string_literal: true
 
 require_relative "./contextual_url"
-require_relative "./xml_processor"
+require_relative "./parsers/event_xml"
 
 module Calendav
   class Event
-    attr_reader :url
+    attr_reader :url, :calendar_data, :etag
 
     def self.from_xml(host, node)
       new(
         ContextualURL.call(host, node.xpath("./dav:href").text),
-        node.xpath(".//dav:prop/*").to_xml,
-        node.namespaces
+        Parsers::EventXML.call(node)
       )
     end
 
-    def initialize(url, attribute_nodes, namespaces = {})
+    def initialize(url, attributes = {})
       @url = url
-      @attribute_nodes = attribute_nodes
-      @namespaces = namespaces
-    end
-
-    def calendar_data
-      attribute_value fragment.xpath("//caldav:calendar-data")
-    end
-
-    def etag
-      attribute_value fragment.xpath("//dav:getetag")
+      @calendar_data = attributes[:calendar_data]
+      @etag = attributes[:etag]
     end
 
     def summary
@@ -34,24 +25,6 @@ module Calendav
     end
 
     private
-
-    attr_reader :attribute_nodes, :namespaces
-
-    def fragment
-      @fragment ||= XMLProcessor.call(
-        "<nodes>#{attribute_nodes}</nodes>", namespaces
-      )
-    end
-
-    def attribute_value(node)
-      return nil if node.children.empty?
-
-      if node.children.any?(&:element?)
-        node.children.select(&:element?).collect(&:to_xml).join
-      else
-        node.children.text
-      end
-    end
 
     def inner_calendar
       Icalendar::Calendar.parse(calendar_data).first