+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## [Unreleased]
 
 * Support creating and deleting calendars.
-* Updating of events.
+* Updating of events (with optional etag check).
 * List all events on a calendar (no timespan required).
 * Updating of calendars.
+* Allow for etag check to be used on event deletions.
+* Finding of single events and calendars by URLs.
 
 ## [0.0.1] - 2021-06-13